split out iterable util
